# Conflicts:
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/com/example/testfirebasedb/activity/DayActivity.java
#	app/src/main/java/com/example/testfirebasedb/activity/DishActivity.java
#	app/src/main/res/layout/activity_dish.xml
#	app/src/main/res/layout/activity_main.xml
#	app/src/main/res/layout/activity_profile.xml
#	app/src/main/res/layout/activity_statistic.xml
#	app/src/main/res/values/colors.xml

+++ app/src/main/java/com/example/testfirebasedb/activity/ExerciseActivity.java
@@ -337,7 +337,7 @@ public class ExerciseActivity extends AppCompatActivity {
             @Override
             public void onDataChange(@NonNull DataSnapshot snapshot) {
                 if (snapshot.exists()) {
-                    int currentCaloOut = snapshot.getValue(Integer.class);
+                    float currentCaloOut = snapshot.getValue(Float.class);
                     float newCaloOut = currentCaloOut + exercise.getCaloBurn();
                     dayRef.child("caloOut").setValue(newCaloOut);
                 }
